Revert "Add link to PyPM package"
This reverts commit 4a600f30d2f221f148de985156b68e4e0421a462.

I prefer not to have this link until we’re very confident that the
package actually works. The hard part in packaging WeasyPrint is with
pycairo, PyGObject 3, and Pango with introspection; not the Python
stuff. PyPM does not list non-Python dependencies and does not seem to
have packages for them.