Use bare commas without spaces for Pango.FontDescription.set_family
Fix a pontential bug reported by Peter Moulder

pjrm: In text.py's call to set_family, note that it's safer
to use ','.join (i.e. without a space).  Some versions of
pango (possibly including the current one) have a bug in
some of the places where a family list is parsed, but using
a plain comma (without space) fixes that.