Fix letter-spacing handling at block boundaries
Pango letter-spacing attribute does not affect the last character in the
text range.  This fix appends a zero-width-space to each TextBox so that
pango letter-spacing will add space following the final character.

Note that an earlier implementation that modified the TextBox width
resulted in increased spacing between boxes but in practice the renderer
often wrapped individual TextBoxes within a single LineBox.