Fix letter-spacing handling at block boundaries
Fix #266 without adding extra chacarcters. The specification is not
clear about how letter-spacing must be implemented, but the browsers
seem to add an extra space after each character. Pango does this too,
but it doesn't do it at the end of the last character. This commit adds
this extra space after the last character.