Use line metrics for struts
That's how it used to work before 412aeb4, but it was replaced by font
metrics for performance issues. Font metrics seems to work well in most
cases, but it fails at least with Ahem for unknown reasons, as the
metrics defined in the font seem pretty good everywhere.

The new code introduced in this commit creates the line (as it was done
before), but doesn't go through the whole line-breaking layout as
there's no text. There's no visible performance loss.