Fix a linebox breaking corner case
Fix #660.

As said in the (very helpful) comment at the beginning of the modified block,
we have to add "child_resume_at" and "initial_skip_stack[1]" skip stacks, but
the old algorithm was obviously wrong when the stacks had more than one level
of depth.

The fix is shorter, easier to read and right (until next bug). A regression
test has been added too.