Discard blocks and tables that don't fit in remaining space
This condition had probably been added a long time ago when layout was done
differently. I don't see why it's there, removing it doesn't break tests and
makes page break before unbreakable blocks that go lower than the end of the
page instead of after.

I may have missed something. If this commit pops out of a git bisect, please
open an issue.

Related to #659 (even if the commit is not related to columns only).