Fix yet another auto table layout
True story: you can read and implement a quite complicated algorithm, use it
for years, and then get an issue related to this algorithm. You can fix the
code, carefully following the algorithm, but still get a bad rendering. Why?
Because the algorithm is wrong. Of course, as it's written by David Baron, it
can't be that wrong: there's actually a red ISSUE block telling that "The way
this describes distribution of widths from column-spanning cells is wrong" and
that I "should refer to the rules for distributing excess width to columns for
intrinsic width calculation".

I did. Guess what: it works.

Fix #685.