Fix tests relying on a modified CSS parsing rule
In the 2014 version of CSS Syntax Module Level 3 [1], meeting an EOF character
before the end of a string would just return the string token. In the 2019
candidate recommendation [2], this behavior must raise a parse error.

This commit fixes a test that was relying on the older behavior.

TinyCSS2 will follow the new specification from version 1.1.0.

[1] https://www.w3.org/TR/2014/CR-css-syntax-3-20140220/#consume-a-string-token
[2] https://www.w3.org/TR/2019/CR-css-syntax-3-20190716/#consume-string-token