tables: add -weasy-break-inside-threshold proprietary extension
break-inside: avoid-page is an all of nothing affair, when applied
on large table row group the result may be less than stellar with
regard to large amounts of useless semi-unintentional white space.

this patch introduces a threshold where the largest size of a
table row group is defined that is to honour the break-inside.
attribute.

usage example:

tbody
{
  break-inside: avoid-page;
  -weasy-break-inside-threshold: 100mm;
}