Fix position of marker box
The marker box is generated by the ::marker pseudo-element of a list item as
the list item’s first child […].

The old behavior was done because of the fuzzy "[With list-style-position:
outside], If the list item is a block container: the marker box is outside the
principal block box […]". Outside here probably means that the box is drawn
outside, not that it's not a child of the list item box.

Related to #945.