Prevent improper var() from crashing with `InvalidValues`
The css validation functions raise `InvalidValues`. Basic CSS validation, done
in Step 2, catches and emits the exception as a WARNING.

In Step 3, when a var() needs to be resolved, those validation functions are
called again. We must catch the `InvalidValues` in compute(), too.

Currently only the `transform` property seems to be endangered.