Don’t crash when improper values are given in var()
When the resolved value of var() is not allowed by the property, we have to get
the parent value (or the initial value if the property is not
inherited). Parent value (and some of the initial values) is already computed,
so we don’t have to call the computer function in this case.

Calling the computer function with already computed values can crash.

Fix #974.