Use start/end instead of left/right to define text alignment offset
As the rtl lineboxes are now right aligned (just like any other block-level
box), we don’t have to take care about direction when calculating offset.

Text is now translated when text-align is different from start. It’s translated
to the right with ltr and to the left with rtl.