Fix a crash with a flexbox corner case
When the flex box height percentage can’t be solved using its parent heihgt,
its computed height is set to 'auto' even if the specified value is not
'auto'. It’s a problem when the flex box has multiple lines, because in this
case we rely on the computed height to distribute the extra space between the
lines.

Fix #1204.