Don’t render table cells as if they are on empty pages
The page_is_empty was probably set to True to force rendering, but using
max_position=inf should already allow all cells to be rendered.

When cells are rendered with page_is_empty set to True, the top margin of their
first children is removed, because margins are always removed at the top of
blocks split between multiple pages.

Fix #1254.