Replace directionality characters with zero-width space
As these characters were removed to avoid crashes when breaking lines, indices
returned were wrong and could cause crashes.

Replacing them with zero-width spaces adds a hack to the hack, but it keeps the
same indicies and should be harmless for rendering.

Of course, the real fix is to have real bidi support.

Fix #1336.