Calculate CSS properties on demand
Instead of calculating all the CSS properties for each element, we only get
them on demand. It saves A LOT of time and memory.

This implementation is just a quick and dirty hack that make tests pass. Even
with this suboptimal code, we get really good results.

Work in progress, has to be improved and cleaned.