Use image id instead of increasing index for image keys
Pillow images are stored in a cache, to be sure that we use the same image when
we have the same URL (and don’t try to retrieve it twice).

Using the image id as a key prevents WeasyPrint to store the image twice in the
PDF, with a large gain of time and size.

Related to #1392.