Include the font hash to create the font name
Using the font hash avoids collisions when we have the same font description
but not the same hash, for example when the same font is used in multiple HTML
instances.

It could be possible to only store the font once, this solution is not optimal
but this more reliable.

Fix #1489.