fix: don't wrap `overflow-wrap: anywhere` unnecessarily
Previously, if `overflow-wrap: anywhere` (or `break-word`) was set on an
inline element (like an `<a>` or `<span>`) whose first word occurred
towards the end of a line, it would have broken that word, even if doing
so wasn't necessary (and wouldn't have happened were it not for the
wrapping element).

This pipes through a `is_line_start` argument to `split_text_box` that
only allows the break on a word if it's the first word in a line.