Explicitly set white background for tests
With the release of Little CMS 2.13, Ghostscript sets a grey background when
generating raster images with no alpha channel. It’s probably a bug in
Ghostscript or in Little CMS, but the specification gives the possibility to
PDF renderers to choose the backdrop color (even if most of them use white).

Setting a white background for tests fixes this problem. We keep the
transparent background in the normal user agent for compatibility purpose, and
users will be able to set a white background if they need a "workaround".