Don’t duplicate the definition of SVG attributes
We want to get this list set only once, because we don’t want to change it
twice each time we add or remove an attribute.

This change probably means that extra work is now done, for style and for
definitions. This should definitely be optimized, but at least it shouldn’t be
wrong.

Tests definitely have to be added for SVG rendering.

Fix #1558.