Improve inheritance of text-decoration-* properties
The text-decoration-* properties are not inherited anymore, they are propagated
according to specific rules.

This commit removes the inheritance, but it doesn’t follow all the rules
defined by the specification. The new behavior is not perfect, but it should be
at least better than the previous code and opens the possibility to fix this
issue correctly in the future.

Failing tests have been added.

Fix #1621.