Follow max-height on footnote area
This solution is by far the cleanest I’ve found to follow max-height without
changing everything. Contrary to my previous tries, it keeps the original style
untouched, doesn’t add dirty properties on the box and only assumes that the
overflowing children are caused by the max-height set on the area. This last
condition is probably true in real-life cases, but we’ll sure find soon that
other users don’t have the real life we expect them to have.

It also has the side effect of following the height attribute too. There’s
nothing in the specification about which properties are allowed in @footnote
rules, only a classical "How would one describe this in the grammar of
CSS3-Page?", so … who cares? We’ll sure find soon that other users actually
care more than the W3C.

The existing tests pass, that’s good news, but we’ll need solid new tests.

Related to #1674.