Don’t shrink split box height when it reaches the bottom of the page
When a box is forced to be displayed (when the page is empty), don’t shrink to
math the bottom of the page. For normal cases the rendering may be better, and
it’s required when elements are forced to be rendered to get their height, for
example during the multicolumn layout.