Don’t use pytest plugins by default
Sometimes, when a new version of pytest, flake8 or isort is released, we get
new problems. Sometimes the bug is quickly fixed, sometimes it ends in drama
(I’m looking at you, tholo/pytest-flake8#87).

Pytest, flake8 and isort are robust tools, and we’d like to continue to use
them and to install them as testing dependencies.

Let’s lauch the tools without Pytest. We keep their configuration in
pyproject.toml so that users who want to use them in their editors can benefit
from the default tweaks. We remove the default pytest options so that packagers
don’t have to edit or remove pyproject.toml, and only have to install
pytest (linting is useless for packaging). We install pytest-xdist to launch
our tests in parallel for the CI.