Use a stable key to store stream fonts
Using a stable key avoids problems that could be caused by different pointers
for the same font face.

Note that this commit doesn’t explicitly avoids two fonts with the same hash in
the generated PDF. But it should be solid enough to handle complex cases, such
as the ones described in #1662.

Fix #1662.