Detect JPEG/MPO images as normal JPEG files
Some cameras store multiple images in JPEG photos, using the MPO format in EXIF
metadata. As Pillow supports this format, it sets the "MPO" format attribute to
the image object, not the "JPEG" one.

This leads to a useless PNG conversion, as MPO are (often, always?) supposed to
be "normal" JPEG files.

See https://github.com/python-pillow/Pillow/issues/1138.

Fix #1777.