Fix bookmarks coordinates
In version 57.x, we used to use two different functions to generate abstract
bookmarks (whose coordinates origin is the top-left corner) and PDF
bookmarks (whose coordinates origin is the bottom-left corner). Now that we
only use one function, Document.make_bookmark_tree, we have to give the scale
and the matrix parameters so that we can retrieve both depending on the
context.

Fix #1815.