Use dictionary to store API options
This commit uses an "option" dictionary to store various API options that were
used as arguments in many public and private functions. This change allows to
easily document default values, to reduce the number of arguments and to avoid
many repetitions in documentation and signatures.

The changes to the public API are minimal, and should only have an impact for
users who passed unnamed arguments.