Don’t duplicate images drawn with multiple dpi ratios
When the same image is displayed twice, with two different sizes, we now only
store the image once, with the maximum size required. Smaller images are then
drawn with a better resolution than expected, but the PDF is obviously smaller
than storing the low-quality version in addition of the high-quality one.

Fix #1877.