Fix WeasyPrint value for input/textarea
Previously we were filling it with a `pydyf.String(None)` when the HTML had no value or even an empty string as value, which was rendering the field with a `None` starting value.

This PR fix this behavior using an empty string on the `/V` if the value is `None`.