Fix "pages" counter in non-root absolute boxes
These boxes were not available through the "descendants" function, because they
are represented by placeholders. They are also not available in the list of
absolute boxes of the page, because they’re not relative to the root page.

The new "placeholders" parameter of "descendants" goes through placeholders and
gets all the absolute children.

Putting "descendants" and "children" in all boxes, not just parents, avoids
useless checks and special cases.

Fix #2029.