Store original and PDF stream images in different cache slots
We used to replace the original image by the PDF stream image in cache after
the first PDF rendering. It’s broken because the PDF stream image is not a
valid image, and can’t be opened by Pillow if it has to be rendered twice.

Replacing the original image with the final one was probably supposed to save
some time and/or some memory, but my quick tests show no performance problem
with this commit.

Fix #1942, fix #2228.