Include floats in calculation of minimum cell height
We calculate the height of a cell by rendering the cell as a block after
setting an auto height to the cell. This way, it can grow if needed.

But setting the height attribute is not enough: we take care of floats in
finish_block_formatting_context, that checks for the style value: we have to
fake this value too.

Related to #2216. Fix #2293.