Don’t include page margin content in content tree
This content could be included in tree as Artifact, but that’s only allowed in PDF 2.0.

This content could also be ignored, but marked content has to be mapped for PDF
validity, and we want to ensure that the marked content list is empty at the end of the
tagging.

PAC and VeraPDF are happy with this solution.

Fix #1836.