Remove useless flag on links and fix structure
Example 3 of chapter 14.8.4.4.2 (Link Elements) of the PDF 1.7 specification explains
what should be done when a paragraph includes a link for example:

- P tag includes Link tag
- Link tag includes OBJR tag

14.7.4.4 (Finding Structure Elements from Content Items) explains that:

- Annotations must have a StructParent

Validators require that the link annotation has to reference a Link tag from the parent
structure (Nums), so that’s why the Link tag is given a dedicated reference in the
parent structure.

It’s still broken for links with nested tags.