Add manual CMap for fonts with missing glyphs
When a glyph is missing in a font, we want to keep the actual glyph id in the PDF
text stream, so that we can map it correctly in ToUnicode. We also want to display a
.notdef glyph instead, using the fallback system described in PDF 2.0 Chapter 9.7.6.3.

The fallback mechanism only works if the glyph is missing from the font. When using the
/Identity-H encoding for the Type 0 font, all glyphs are mapped, and no glyph is marked
as missing. So, we have to build a CMap with only used glyphs in order to have missing
glyphs actually missing from the Type 0 font and replaced by .notdef.