Possibly include child’s bottom margin to find earlier page break
When a child’s content fits but not its bottom decorations, we have to re-render the
child removing the size of its bottom decorations so that it can find an earlier break.
These decorations always include bottom padding and border.

When the parent’s bottom margin doesn’t collapse, we also have to include the child’s
bottom margin, as it will also increase the parent’s size.