Add font-face support to SVG
Not sure if this is how it should be implemented, but this fixes the
issue where `@font-face` declarations within a SVG `<style>` where not
parsed (and thus the font was not available for the SVG).

It does add the font to the font_config, which means the font will be
available to the rest of the document, however this seems to be the case
in browsers as well.